requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
httpx>=0.27.0
APScheduler>=3.10.0
matplotlib>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0
chromadb>=0.4
scikit-learn>=1.3
joblib>=1.3
